Add files via upload
Actually this code was one of exercise which I couldn't get full mark. I appreciate any help, to complete it. Here is the question. note that Sindekht and Irsan are just names.

One day Sindekht and Irsa were discussing the price of laptops and their quality. Sindekht guesses that the more expensive a laptop is, the better its quality. But Irsa claims that he can find 2 laptops, the price of the first one is lower than the second one, but the quality is higher than the second one, and he can reject Sindekht's guess. Now you have to write a program to help Irsa check her claim.

The specifications of the laptop will be given to you. The first line of the input contains the number n, which indicates the number of laptops. Each of the next n lines contains two numbers, the first number shows the price of the laptop and the second number shows the quality of that laptop. If you can find two laptops that meet the conditions mentioned by Irsa, print happy irsa, otherwise, print poor irsa (please pay attention to the small letters, all letters are written in lower case.)